requirements.txt
asgiref==3.5.2
boto3==1.24.67
botocore==1.27.67
dj-database-url==1.0.0
Django==4.1
django-archive==0.2.0
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
